requirements-auth.txt
Flask-Login==0.6.3
Flask-WTF==1.2.1
